+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/playlist/HlsPlaylistParser.java
@@ -335,6 +335,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
   public HlsPlaylist parse(Uri uri, InputStream inputStream) throws IOException {
     BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
     Queue<String> extraLines = new ArrayDeque<>();
+    MatcherCache matcherCache = new MatcherCache();
     String line;
     try {
       if (!checkPlaylistHeader(reader)) {
@@ -347,7 +348,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           // Do nothing.
         } else if (line.startsWith(TAG_STREAM_INF)) {
           extraLines.add(line);
-          return parseMultivariantPlaylist(new LineIterator(extraLines, reader), uri);
+          return parseMultivariantPlaylist(new LineIterator(extraLines, reader), uri, matcherCache);
         } else if (line.startsWith(TAG_TARGET_DURATION)
             || line.startsWith(TAG_MEDIA_SEQUENCE)
             || line.startsWith(TAG_MEDIA_DURATION)
@@ -361,7 +362,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
               multivariantPlaylist,
               previousMediaPlaylist,
               new LineIterator(extraLines, reader),
-              uri);
+              uri,
+              matcherCache);
         } else {
           extraLines.add(line);
         }
@@ -427,7 +429,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
   }
 
   private static HlsMultivariantPlaylist parseMultivariantPlaylist(
-      LineIterator iterator, Uri playlistUri) throws IOException {
+      LineIterator iterator, Uri playlistUri, MatcherCache matcherCache) throws IOException {
+    String baseUri = playlistUri.toString();
     HashMap<Uri, ArrayList<VariantInfo>> urlToVariantInfos = new HashMap<>();
     HashMap<String, String> variableDefinitions = new HashMap<>();
     ArrayList<Variant> variants = new ArrayList<>();
@@ -455,14 +458,16 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
 
       if (line.startsWith(TAG_DEFINE)) {
         @Nullable
-        String variableName = parseOptionalStringAttr(line, REGEX_NAME, variableDefinitions);
+        String variableName =
+            parseOptionalStringAttr(line, REGEX_NAME, variableDefinitions, matcherCache);
         if (variableName != null) {
           verifyVariableNameNotContainedOrThrow(variableName, variableDefinitions);
           variableDefinitions.put(
               /* key= */ variableName,
-              /* value= */ parseStringAttr(line, REGEX_VALUE, variableDefinitions));
+              /* value= */ parseStringAttr(line, REGEX_VALUE, variableDefinitions, matcherCache));
         } else {
-          variableName = parseStringAttr(line, REGEX_QUERY_PARAM, variableDefinitions);
+          variableName =
+              parseStringAttr(line, REGEX_QUERY_PARAM, variableDefinitions, matcherCache);
           verifyVariableNameNotContainedOrThrow(variableName, variableDefinitions);
           @Nullable String value = playlistUri.getQueryParameter(variableName);
           if (value == null) {
@@ -479,22 +484,27 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         mediaTags.add(line);
       } else if (line.startsWith(TAG_SESSION_KEY)) {
         String keyFormat =
-            parseOptionalStringAttr(line, REGEX_KEYFORMAT, KEYFORMAT_IDENTITY, variableDefinitions);
-        SchemeData schemeData = parseDrmSchemeData(line, keyFormat, variableDefinitions);
+            parseOptionalStringAttr(
+                line, REGEX_KEYFORMAT, KEYFORMAT_IDENTITY, variableDefinitions, matcherCache);
+        SchemeData schemeData =
+            parseDrmSchemeData(line, keyFormat, variableDefinitions, matcherCache);
         if (schemeData != null) {
-          String method = parseStringAttr(line, REGEX_METHOD, variableDefinitions);
+          String method = parseStringAttr(line, REGEX_METHOD, variableDefinitions, matcherCache);
           String scheme = parseEncryptionScheme(method);
           sessionKeyDrmInitData.add(new DrmInitData(scheme, schemeData));
         }
       } else if (line.startsWith(TAG_STREAM_INF) || isIFrameOnlyVariant) {
         noClosedCaptions |= line.contains(ATTR_CLOSED_CAPTIONS_NONE);
         int roleFlags = isIFrameOnlyVariant ? C.ROLE_FLAG_TRICK_PLAY : 0;
-        int peakBitrate = parseIntAttr(line, REGEX_BANDWIDTH);
-        int averageBitrate = parseOptionalIntAttr(line, REGEX_AVERAGE_BANDWIDTH, -1);
-        String videoRange = parseOptionalStringAttr(line, REGEX_VIDEO_RANGE, variableDefinitions);
-        String codecs = parseOptionalStringAttr(line, REGEX_CODECS, variableDefinitions);
+        int peakBitrate = parseIntAttr(line, REGEX_BANDWIDTH, matcherCache);
+        int averageBitrate = parseOptionalIntAttr(line, REGEX_AVERAGE_BANDWIDTH, -1, matcherCache);
+        String videoRange =
+            parseOptionalStringAttr(line, REGEX_VIDEO_RANGE, variableDefinitions, matcherCache);
+        String codecs =
+            parseOptionalStringAttr(line, REGEX_CODECS, variableDefinitions, matcherCache);
         String supplementalCodecsStrings =
-            parseOptionalStringAttr(line, REGEX_SUPPLEMENTAL_CODECS, variableDefinitions);
+            parseOptionalStringAttr(
+                line, REGEX_SUPPLEMENTAL_CODECS, variableDefinitions, matcherCache);
         String supplementalCodecs = null;
         String supplementalProfiles = null; // i.e. Compatibility brand
         if (supplementalCodecsStrings != null) {
@@ -515,7 +525,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         }
 
         String resolutionString =
-            parseOptionalStringAttr(line, REGEX_RESOLUTION, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_RESOLUTION, variableDefinitions, matcherCache);
         int width;
         int height;
         if (resolutionString != null) {
@@ -533,33 +543,37 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         }
         float frameRate = Format.NO_VALUE;
         String frameRateString =
-            parseOptionalStringAttr(line, REGEX_FRAME_RATE, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_FRAME_RATE, variableDefinitions, matcherCache);
         if (frameRateString != null) {
           frameRate = Float.parseFloat(frameRateString);
         }
         @Nullable
-        String pathwayId = parseOptionalStringAttr(line, REGEX_PATHWAY_ID, variableDefinitions);
-        String videoGroupId = parseOptionalStringAttr(line, REGEX_VIDEO, variableDefinitions);
-        String audioGroupId = parseOptionalStringAttr(line, REGEX_AUDIO, variableDefinitions);
+        String pathwayId =
+            parseOptionalStringAttr(line, REGEX_PATHWAY_ID, variableDefinitions, matcherCache);
+        String videoGroupId =
+            parseOptionalStringAttr(line, REGEX_VIDEO, variableDefinitions, matcherCache);
+        String audioGroupId =
+            parseOptionalStringAttr(line, REGEX_AUDIO, variableDefinitions, matcherCache);
         String subtitlesGroupId =
-            parseOptionalStringAttr(line, REGEX_SUBTITLES, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_SUBTITLES, variableDefinitions, matcherCache);
         String closedCaptionsGroupId =
-            parseOptionalStringAttr(line, REGEX_CLOSED_CAPTIONS, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_CLOSED_CAPTIONS, variableDefinitions, matcherCache);
         @Nullable
         String stableVariantId =
-            parseOptionalStringAttr(line, REGEX_STABLE_VARIANT_ID, variableDefinitions);
+            parseOptionalStringAttr(
+                line, REGEX_STABLE_VARIANT_ID, variableDefinitions, matcherCache);
         Uri uri;
         if (isIFrameOnlyVariant) {
           uri =
               UriUtil.resolveToUri(
-                  playlistUri.toString(), parseStringAttr(line, REGEX_URI, variableDefinitions));
+                  baseUri, parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache));
         } else if (!iterator.hasNext()) {
           throw ParserException.createForMalformedManifest(
               "#EXT-X-STREAM-INF must be followed by another line", /* cause= */ null);
         } else {
           // The following line contains #EXT-X-STREAM-INF's URI.
-          line = replaceVariableReferences(iterator.next(), variableDefinitions);
-          uri = UriUtil.resolveToUri(playlistUri.toString(), line);
+          line = replaceVariableReferences(iterator.next(), variableDefinitions, matcherCache);
+          uri = UriUtil.resolveToUri(baseUri, line);
         }
 
         Format format =
@@ -621,29 +635,29 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
 
     for (int i = 0; i < mediaTags.size(); i++) {
       line = mediaTags.get(i);
-      String groupId = parseStringAttr(line, REGEX_GROUP_ID, variableDefinitions);
-      String name = parseStringAttr(line, REGEX_NAME, variableDefinitions);
+      String groupId = parseStringAttr(line, REGEX_GROUP_ID, variableDefinitions, matcherCache);
+      String name = parseStringAttr(line, REGEX_NAME, variableDefinitions, matcherCache);
       @Nullable
       String stableRenditionId =
-          parseOptionalStringAttr(line, REGEX_STABLE_RENDITION_ID, variableDefinitions);
+          parseOptionalStringAttr(
+              line, REGEX_STABLE_RENDITION_ID, variableDefinitions, matcherCache);
       Format.Builder formatBuilder =
           new Format.Builder()
               .setId(groupId + ":" + name)
               .setLabel(name)
               .setContainerMimeType(MimeTypes.APPLICATION_M3U8)
-              .setSelectionFlags(parseSelectionFlags(line))
-              .setRoleFlags(parseRoleFlags(line, variableDefinitions))
-              .setLanguage(parseOptionalStringAttr(line, REGEX_LANGUAGE, variableDefinitions));
+              .setSelectionFlags(parseSelectionFlags(line, matcherCache))
+              .setRoleFlags(parseRoleFlags(line, variableDefinitions, matcherCache))
+              .setLanguage(
+                  parseOptionalStringAttr(line, REGEX_LANGUAGE, variableDefinitions, matcherCache));
 
-      @Nullable String referenceUri = parseOptionalStringAttr(line, REGEX_URI, variableDefinitions);
       @Nullable
-      Uri uri =
-          referenceUri == null
-              ? null
-              : UriUtil.resolveToUri(/* baseUri= */ playlistUri.toString(), referenceUri);
+      String referenceUri =
+          parseOptionalStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
+      @Nullable Uri uri = referenceUri == null ? null : UriUtil.resolveToUri(baseUri, referenceUri);
       Metadata metadata =
           new Metadata(new HlsTrackMetadataEntry(groupId, name, Collections.emptyList()));
-      switch (parseStringAttr(line, REGEX_TYPE, variableDefinitions)) {
+      switch (parseStringAttr(line, REGEX_TYPE, variableDefinitions, matcherCache)) {
         case TYPE_VIDEO:
           @Nullable Variant variant = getVariantWithVideoGroup(variants, groupId);
           if (variant != null) {
@@ -675,7 +689,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           }
           @Nullable
           String channelsString =
-              parseOptionalStringAttr(line, REGEX_CHANNELS, variableDefinitions);
+              parseOptionalStringAttr(line, REGEX_CHANNELS, variableDefinitions, matcherCache);
           if (channelsString != null) {
             int channelCount = Integer.parseInt(Util.splitAtFirst(channelsString, "/")[0]);
             formatBuilder.setChannelCount(channelCount);
@@ -714,7 +728,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           }
           break;
         case TYPE_CLOSED_CAPTIONS:
-          String instreamId = parseStringAttr(line, REGEX_INSTREAM_ID, variableDefinitions);
+          String instreamId =
+              parseStringAttr(line, REGEX_INSTREAM_ID, variableDefinitions, matcherCache);
           int accessibilityChannel;
           if (instreamId.startsWith("CC")) {
             sampleMimeType = MimeTypes.APPLICATION_CEA608;
@@ -794,8 +809,10 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
       HlsMultivariantPlaylist multivariantPlaylist,
       @Nullable HlsMediaPlaylist previousMediaPlaylist,
       LineIterator iterator,
-      Uri playlistUri)
+      Uri playlistUri,
+      MatcherCache matcherCache)
       throws IOException {
+    String baseUri = playlistUri.toString();
     @HlsMediaPlaylist.PlaylistType int playlistType = HlsMediaPlaylist.PLAYLIST_TYPE_UNKNOWN;
     long startOffsetUs = C.TIME_UNSET;
     long mediaSequence = 0;
@@ -854,7 +871,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
       }
 
       if (line.startsWith(TAG_PLAYLIST_TYPE)) {
-        String playlistTypeString = parseStringAttr(line, REGEX_PLAYLIST_TYPE, variableDefinitions);
+        String playlistTypeString =
+            parseStringAttr(line, REGEX_PLAYLIST_TYPE, variableDefinitions, matcherCache);
         if ("VOD".equals(playlistTypeString)) {
           playlistType = HlsMediaPlaylist.PLAYLIST_TYPE_VOD;
         } else if ("EVENT".equals(playlistTypeString)) {
@@ -863,17 +881,21 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
       } else if (line.equals(TAG_IFRAME)) {
         isIFrameOnly = true;
       } else if (line.startsWith(TAG_START)) {
-        startOffsetUs = (long) (parseDoubleAttr(line, REGEX_TIME_OFFSET) * C.MICROS_PER_SECOND);
+        startOffsetUs =
+            (long) (parseDoubleAttr(line, REGEX_TIME_OFFSET, matcherCache) * C.MICROS_PER_SECOND);
         preciseStart =
-            parseOptionalBooleanAttribute(line, REGEX_PRECISE, /* defaultValue= */ false);
+            parseOptionalBooleanAttribute(
+                line, REGEX_PRECISE, /* defaultValue= */ false, matcherCache);
       } else if (line.startsWith(TAG_SERVER_CONTROL)) {
-        serverControl = parseServerControl(line);
+        serverControl = parseServerControl(line, matcherCache);
       } else if (line.startsWith(TAG_PART_INF)) {
-        double partTargetDurationSeconds = parseDoubleAttr(line, REGEX_PART_TARGET_DURATION);
+        double partTargetDurationSeconds =
+            parseDoubleAttr(line, REGEX_PART_TARGET_DURATION, matcherCache);
         partTargetDurationUs = (long) (partTargetDurationSeconds * C.MICROS_PER_SECOND);
       } else if (line.startsWith(TAG_INIT_SEGMENT)) {
-        String uri = parseStringAttr(line, REGEX_URI, variableDefinitions);
-        String byteRange = parseOptionalStringAttr(line, REGEX_ATTR_BYTERANGE, variableDefinitions);
+        String uri = parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
+        String byteRange =
+            parseOptionalStringAttr(line, REGEX_ATTR_BYTERANGE, variableDefinitions, matcherCache);
         if (byteRange != null) {
           String[] splitByteRange = Util.split(byteRange, "@");
           segmentByteRangeLength = Long.parseLong(splitByteRange[0]);
@@ -904,19 +926,24 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         }
         segmentByteRangeLength = C.LENGTH_UNSET;
       } else if (line.startsWith(TAG_TARGET_DURATION)) {
-        targetDurationUs = parseIntAttr(line, REGEX_TARGET_DURATION) * C.MICROS_PER_SECOND;
+        targetDurationUs =
+            parseIntAttr(line, REGEX_TARGET_DURATION, matcherCache) * C.MICROS_PER_SECOND;
       } else if (line.startsWith(TAG_MEDIA_SEQUENCE)) {
-        mediaSequence = parseLongAttr(line, REGEX_MEDIA_SEQUENCE);
+        mediaSequence = parseLongAttr(line, REGEX_MEDIA_SEQUENCE, matcherCache);
         segmentMediaSequence = mediaSequence;
       } else if (line.startsWith(TAG_VERSION)) {
-        version = parseIntAttr(line, REGEX_VERSION);
+        version = parseIntAttr(line, REGEX_VERSION, matcherCache);
       } else if (line.startsWith(TAG_DEFINE)) {
-        String variableName = parseOptionalStringAttr(line, REGEX_NAME, variableDefinitions);
-        String queryParam = parseOptionalStringAttr(line, REGEX_QUERY_PARAM, variableDefinitions);
+        @Nullable
+        String variableName =
+            parseOptionalStringAttr(line, REGEX_NAME, variableDefinitions, matcherCache);
+        @Nullable
+        String queryParam =
+            parseOptionalStringAttr(line, REGEX_QUERY_PARAM, variableDefinitions, matcherCache);
         if (variableName != null) {
           verifyVariableNameNotContainedOrThrow(variableName, variableDefinitions);
           variableDefinitions.put(
-              variableName, parseStringAttr(line, REGEX_VALUE, variableDefinitions));
+              variableName, parseStringAttr(line, REGEX_VALUE, variableDefinitions, matcherCache));
         } else if (queryParam != null) {
           verifyVariableNameNotContainedOrThrow(queryParam, variableDefinitions);
           @Nullable String value = playlistUri.getQueryParameter(queryParam);
@@ -926,7 +953,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           }
           variableDefinitions.put(queryParam, value);
         } else {
-          String importName = parseStringAttr(line, REGEX_IMPORT, variableDefinitions);
+          String importName =
+              parseStringAttr(line, REGEX_IMPORT, variableDefinitions, matcherCache);
           verifyVariableNameNotContainedOrThrow(importName, variableDefinitions);
           String value = multivariantPlaylist.variableDefinitions.get(importName);
           if (value != null) {
@@ -936,10 +964,11 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           }
         }
       } else if (line.startsWith(TAG_MEDIA_DURATION)) {
-        segmentDurationUs = parseTimeSecondsToUs(line, REGEX_MEDIA_DURATION);
-        segmentTitle = parseOptionalStringAttr(line, REGEX_MEDIA_TITLE, "", variableDefinitions);
+        segmentDurationUs = parseTimeSecondsToUs(line, REGEX_MEDIA_DURATION, matcherCache);
+        segmentTitle =
+            parseOptionalStringAttr(line, REGEX_MEDIA_TITLE, "", variableDefinitions, matcherCache);
       } else if (line.startsWith(TAG_SKIP)) {
-        int skippedSegmentCount = parseIntAttr(line, REGEX_SKIPPED_SEGMENTS);
+        int skippedSegmentCount = parseIntAttr(line, REGEX_SKIPPED_SEGMENTS, matcherCache);
         checkState(previousMediaPlaylist != null && segments.isEmpty());
         int startIndex = (int) (mediaSequence - castNonNull(previousMediaPlaylist).mediaSequence);
         int endIndex = startIndex + skippedSegmentCount;
@@ -976,20 +1005,23 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           segmentMediaSequence++;
         }
       } else if (line.startsWith(TAG_KEY)) {
-        String method = parseStringAttr(line, REGEX_METHOD, variableDefinitions);
+        String method = parseStringAttr(line, REGEX_METHOD, variableDefinitions, matcherCache);
         String keyFormat =
-            parseOptionalStringAttr(line, REGEX_KEYFORMAT, KEYFORMAT_IDENTITY, variableDefinitions);
+            parseOptionalStringAttr(
+                line, REGEX_KEYFORMAT, KEYFORMAT_IDENTITY, variableDefinitions, matcherCache);
         fullSegmentEncryptionKeyUri = null;
         fullSegmentEncryptionIV = null;
         if (METHOD_NONE.equals(method)) {
           currentSchemeDatas.clear();
           cachedDrmInitData = null;
         } else /* !METHOD_NONE.equals(method) */ {
-          fullSegmentEncryptionIV = parseOptionalStringAttr(line, REGEX_IV, variableDefinitions);
+          fullSegmentEncryptionIV =
+              parseOptionalStringAttr(line, REGEX_IV, variableDefinitions, matcherCache);
           if (KEYFORMAT_IDENTITY.equals(keyFormat)) {
             if (METHOD_AES_128.equals(method)) {
               // The segment is fully encrypted using an identity key.
-              fullSegmentEncryptionKeyUri = parseStringAttr(line, REGEX_URI, variableDefinitions);
+              fullSegmentEncryptionKeyUri =
+                  parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
             } else {
               // Do nothing. Samples are encrypted using an identity key, but this is not supported.
               // Hopefully, a traditional DRM alternative is also provided.
@@ -998,7 +1030,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
             if (encryptionScheme == null) {
               encryptionScheme = parseEncryptionScheme(method);
             }
-            SchemeData schemeData = parseDrmSchemeData(line, keyFormat, variableDefinitions);
+            SchemeData schemeData =
+                parseDrmSchemeData(line, keyFormat, variableDefinitions, matcherCache);
             if (schemeData != null) {
               cachedDrmInitData = null;
               currentSchemeDatas.put(keyFormat, schemeData);
@@ -1006,7 +1039,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           }
         }
       } else if (line.startsWith(TAG_BYTERANGE)) {
-        String byteRange = parseStringAttr(line, REGEX_BYTERANGE, variableDefinitions);
+        String byteRange =
+            parseStringAttr(line, REGEX_BYTERANGE, variableDefinitions, matcherCache);
         String[] splitByteRange = Util.split(byteRange, "@");
         segmentByteRangeLength = Long.parseLong(splitByteRange[0]);
         if (splitByteRange.length > 1) {
@@ -1030,27 +1064,30 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
       } else if (line.equals(TAG_ENDLIST)) {
         hasEndTag = true;
       } else if (line.startsWith(TAG_RENDITION_REPORT)) {
-        long lastMediaSequence = parseOptionalLongAttr(line, REGEX_LAST_MSN, C.INDEX_UNSET);
-        int lastPartIndex = parseOptionalIntAttr(line, REGEX_LAST_PART, C.INDEX_UNSET);
-        String uri = parseStringAttr(line, REGEX_URI, variableDefinitions);
+        long lastMediaSequence =
+            parseOptionalLongAttr(line, REGEX_LAST_MSN, C.INDEX_UNSET, matcherCache);
+        int lastPartIndex =
+            parseOptionalIntAttr(line, REGEX_LAST_PART, C.INDEX_UNSET, matcherCache);
+        String uri = parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
+        Uri renditionReportUri = Uri.parse(UriUtil.resolve(baseUri, uri));
         renditionReports.add(
-            new RenditionReport(
-                Uri.parse(UriUtil.resolve(/* baseUri= */ playlistUri.toString(), uri)),
-                lastMediaSequence,
-                lastPartIndex));
+            new RenditionReport(renditionReportUri, lastMediaSequence, lastPartIndex));
       } else if (line.startsWith(TAG_PRELOAD_HINT)) {
         if (preloadPart != null) {
           continue;
         }
-        String type = parseStringAttr(line, REGEX_PRELOAD_HINT_TYPE, variableDefinitions);
+        String type =
+            parseStringAttr(line, REGEX_PRELOAD_HINT_TYPE, variableDefinitions, matcherCache);
         if (!TYPE_PART.equals(type)) {
           continue;
         }
-        String url = parseStringAttr(line, REGEX_URI, variableDefinitions);
+        String url = parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
         long byteRangeStart =
-            parseOptionalLongAttr(line, REGEX_BYTERANGE_START, /* defaultValue= */ C.LENGTH_UNSET);
+            parseOptionalLongAttr(
+                line, REGEX_BYTERANGE_START, /* defaultValue= */ C.LENGTH_UNSET, matcherCache);
         long byteRangeLength =
-            parseOptionalLongAttr(line, REGEX_BYTERANGE_LENGTH, /* defaultValue= */ C.LENGTH_UNSET);
+            parseOptionalLongAttr(
+                line, REGEX_BYTERANGE_LENGTH, /* defaultValue= */ C.LENGTH_UNSET, matcherCache);
         @Nullable
         String segmentEncryptionIV =
             getSegmentEncryptionIV(
@@ -1085,16 +1122,19 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         String segmentEncryptionIV =
             getSegmentEncryptionIV(
                 segmentMediaSequence, fullSegmentEncryptionKeyUri, fullSegmentEncryptionIV);
-        String url = parseStringAttr(line, REGEX_URI, variableDefinitions);
+        String url = parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
         long partDurationUs =
-            (long) (parseDoubleAttr(line, REGEX_ATTR_DURATION) * C.MICROS_PER_SECOND);
+            (long) (parseDoubleAttr(line, REGEX_ATTR_DURATION, matcherCache) * C.MICROS_PER_SECOND);
         boolean isIndependent =
-            parseOptionalBooleanAttribute(line, REGEX_INDEPENDENT, /* defaultValue= */ false);
+            parseOptionalBooleanAttribute(
+                line, REGEX_INDEPENDENT, /* defaultValue= */ false, matcherCache);
         // The first part of a segment is always independent if the segments are independent.
         isIndependent |= hasIndependentSegmentsTag && trailingParts.isEmpty();
-        boolean isGap = parseOptionalBooleanAttribute(line, REGEX_GAP, /* defaultValue= */ false);
+        boolean isGap =
+            parseOptionalBooleanAttribute(line, REGEX_GAP, /* defaultValue= */ false, matcherCache);
         @Nullable
-        String byteRange = parseOptionalStringAttr(line, REGEX_ATTR_BYTERANGE, variableDefinitions);
+        String byteRange =
+            parseOptionalStringAttr(line, REGEX_ATTR_BYTERANGE, variableDefinitions, matcherCache);
         long partByteRangeLength = C.LENGTH_UNSET;
         if (byteRange != null) {
           String[] splitByteRange = Util.split(byteRange, "@");
@@ -1133,36 +1173,39 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
           partByteRangeOffset += partByteRangeLength;
         }
       } else if (line.startsWith(TAG_DATERANGE)
-          && parseOptionalStringAttr(line, REGEX_CLASS, /* defaultValue= */ "", variableDefinitions)
+          && parseOptionalStringAttr(
+                  line, REGEX_CLASS, /* defaultValue= */ "", variableDefinitions, matcherCache)
               .equals(DATERANGE_CLASS_INTERSTITIALS)) {
-        String id = parseStringAttr(line, REGEX_ID, variableDefinitions);
+        String id = parseStringAttr(line, REGEX_ID, variableDefinitions, matcherCache);
         @Nullable Uri assetUri = null;
-        String assetUriString = parseOptionalStringAttr(line, REGEX_ASSET_URI, variableDefinitions);
+        String assetUriString =
+            parseOptionalStringAttr(line, REGEX_ASSET_URI, variableDefinitions, matcherCache);
         if (assetUriString != null) {
           assetUri = Uri.parse(assetUriString);
         }
         @Nullable Uri assetListUri = null;
         String assetListUriString =
-            parseOptionalStringAttr(line, REGEX_ASSET_LIST_URI, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_ASSET_LIST_URI, variableDefinitions, matcherCache);
         if (assetListUriString != null) {
           assetListUri = Uri.parse(assetListUriString);
         }
         long startDateUnixUs = C.TIME_UNSET;
         @Nullable
         String startDateUnixMsString =
-            parseOptionalStringAttr(line, REGEX_START_DATE, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_START_DATE, variableDefinitions, matcherCache);
         if (startDateUnixMsString != null) {
           startDateUnixUs = msToUs(parseXsDateTime(startDateUnixMsString));
         }
         long endDateUnixUs = C.TIME_UNSET;
         @Nullable
         String endDateUnixMsString =
-            parseOptionalStringAttr(line, REGEX_END_DATE, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_END_DATE, variableDefinitions, matcherCache);
         if (endDateUnixMsString != null) {
           endDateUnixUs = msToUs(parseXsDateTime(endDateUnixMsString));
         }
         List<@Interstitial.CueTriggerType String> cue = new ArrayList<>();
-        String cueString = parseOptionalStringAttr(line, REGEX_CUE, variableDefinitions);
+        String cueString =
+            parseOptionalStringAttr(line, REGEX_CUE, variableDefinitions, matcherCache);
         if (cueString != null) {
           String[] identifiers = Util.split(/* value= */ cueString, /* regex= */ ",");
           for (String identifier : identifiers) {
@@ -1178,30 +1221,35 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
             }
           }
         }
-        double durationSec = parseOptionalDoubleAttr(line, REGEX_ATTR_DURATION_PREFIXED, -1.0d);
+        double durationSec =
+            parseOptionalDoubleAttr(line, REGEX_ATTR_DURATION_PREFIXED, -1.0d, matcherCache);
         long durationUs = C.TIME_UNSET;
         if (durationSec >= 0) {
           durationUs = (long) (durationSec * C.MICROS_PER_SECOND);
         }
-        double plannedDurationSec = parseOptionalDoubleAttr(line, REGEX_PLANNED_DURATION, -1.0d);
+        double plannedDurationSec =
+            parseOptionalDoubleAttr(line, REGEX_PLANNED_DURATION, -1.0d, matcherCache);
         long plannedDurationUs = C.TIME_UNSET;
         if (plannedDurationSec >= 0) {
           plannedDurationUs = (long) (plannedDurationSec * C.MICROS_PER_SECOND);
         }
-        boolean endOnNext = parseOptionalBooleanAttribute(line, REGEX_END_ON_NEXT, false);
+        boolean endOnNext =
+            parseOptionalBooleanAttribute(line, REGEX_END_ON_NEXT, false, matcherCache);
         double resumeOffsetUsDouble =
-            parseOptionalDoubleAttr(line, REGEX_RESUME_OFFSET, Double.MIN_VALUE);
+            parseOptionalDoubleAttr(line, REGEX_RESUME_OFFSET, Double.MIN_VALUE, matcherCache);
         long resumeOffsetUs = C.TIME_UNSET;
         if (resumeOffsetUsDouble != Double.MIN_VALUE) {
           resumeOffsetUs = (long) (resumeOffsetUsDouble * C.MICROS_PER_SECOND);
         }
-        double playoutLimitSec = parseOptionalDoubleAttr(line, REGEX_PLAYOUT_LIMIT, -1.0d);
+        double playoutLimitSec =
+            parseOptionalDoubleAttr(line, REGEX_PLAYOUT_LIMIT, -1.0d, matcherCache);
         long playoutLimitUs = C.TIME_UNSET;
         if (playoutLimitSec >= 0) {
           playoutLimitUs = (long) (playoutLimitSec * C.MICROS_PER_SECOND);
         }
         List<@Interstitial.SnapType String> snapTypes = new ArrayList<>();
-        String snapTypesString = parseOptionalStringAttr(line, REGEX_SNAP, variableDefinitions);
+        String snapTypesString =
+            parseOptionalStringAttr(line, REGEX_SNAP, variableDefinitions, matcherCache);
         if (snapTypesString != null) {
           String[] snapTypesSplit = Util.split(snapTypesString, ",");
           for (String snapType : snapTypesSplit) {
@@ -1218,7 +1266,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         }
         List<@Interstitial.NavigationRestriction String> restrictions = new ArrayList<>();
         String restrictionsString =
-            parseOptionalStringAttr(line, REGEX_RESTRICT, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_RESTRICT, variableDefinitions, matcherCache);
         if (restrictionsString != null) {
           String[] restrictionsSplit = Util.split(restrictionsString, ",");
           for (String restriction : restrictionsSplit) {
@@ -1236,14 +1284,16 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
 
         @Nullable Boolean contentMayVary = null;
         String contentMayVaryString =
-            parseOptionalStringAttr(line, REGEX_CONTENT_MAY_VARY, variableDefinitions);
+            parseOptionalStringAttr(
+                line, REGEX_CONTENT_MAY_VARY, variableDefinitions, matcherCache);
         if (contentMayVaryString != null) {
           contentMayVary = !contentMayVaryString.equals(BOOLEAN_FALSE); // default is true
         }
 
         @Nullable String timelineOccupies = null;
         String timelineOccupiesString =
-            parseOptionalStringAttr(line, REGEX_TIMELINE_OCCUPIES, variableDefinitions);
+            parseOptionalStringAttr(
+                line, REGEX_TIMELINE_OCCUPIES, variableDefinitions, matcherCache);
         if (timelineOccupiesString != null) {
           if (timelineOccupiesString.equals(Interstitial.TIMELINE_OCCUPIES_RANGE)) {
             timelineOccupies = Interstitial.TIMELINE_OCCUPIES_RANGE;
@@ -1254,7 +1304,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
 
         @Nullable String timelineStyle = null;
         String timelineStyleString =
-            parseOptionalStringAttr(line, REGEX_TIMELINE_STYLE, variableDefinitions);
+            parseOptionalStringAttr(line, REGEX_TIMELINE_STYLE, variableDefinitions, matcherCache);
         if (timelineStyleString != null) {
           if (timelineStyleString.equals(Interstitial.TIMELINE_STYLE_PRIMARY)) {
             timelineStyle = Interstitial.TIMELINE_STYLE_PRIMARY;
@@ -1264,24 +1314,26 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         }
 
         double skipControlOffsetSec =
-            parseOptionalDoubleAttr(line, REGEX_SKIP_CONTROL_OFFSET, -1.0d);
+            parseOptionalDoubleAttr(line, REGEX_SKIP_CONTROL_OFFSET, -1.0d, matcherCache);
         long skipControlOffsetUs = C.TIME_UNSET;
         if (skipControlOffsetSec >= 0) {
           skipControlOffsetUs = (long) (skipControlOffsetSec * C.MICROS_PER_SECOND);
         }
         double skipControlDurationSec =
-            parseOptionalDoubleAttr(line, REGEX_SKIP_CONTROL_DURATION, -1.0d);
+            parseOptionalDoubleAttr(line, REGEX_SKIP_CONTROL_DURATION, -1.0d, matcherCache);
         long skipControlDurationUs = C.TIME_UNSET;
         if (skipControlDurationSec >= 0) {
           skipControlDurationUs = (long) (skipControlDurationSec * C.MICROS_PER_SECOND);
         }
         @Nullable
         String skipControlLabelId =
-            parseOptionalStringAttr(line, REGEX_SKIP_CONTROL_LABEL_ID, variableDefinitions);
+            parseOptionalStringAttr(
+                line, REGEX_SKIP_CONTROL_LABEL_ID, variableDefinitions, matcherCache);
 
         List<HlsMediaPlaylist.ClientDefinedAttribute> clientDefinedAttributes = new ArrayList<>();
         String attributes = line.substring("#EXT-X-DATERANGE:".length());
-        Matcher matcher = REGEX_CLIENT_DEFINED_ATTRIBUTE_PREFIX.matcher(attributes);
+        Matcher matcher =
+            matcherCache.obtainMatcher(REGEX_CLIENT_DEFINED_ATTRIBUTE_PREFIX, attributes);
         while (matcher.find()) {
           String attributePrefix = matcher.group();
           switch (attributePrefix) {
@@ -1304,7 +1356,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
                   parseClientDefinedAttribute(
                       attributes,
                       attributePrefix.substring(0, attributePrefix.length() - 1),
-                      variableDefinitions));
+                      variableDefinitions,
+                      matcherCache));
               break;
           }
         }
@@ -1339,7 +1392,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
             getSegmentEncryptionIV(
                 segmentMediaSequence, fullSegmentEncryptionKeyUri, fullSegmentEncryptionIV);
         segmentMediaSequence++;
-        String segmentUri = replaceVariableReferences(line, variableDefinitions);
+        String segmentUri = replaceVariableReferences(line, variableDefinitions, matcherCache);
         @Nullable Segment inferredInitSegment = urlToInferredInitSegment.get(segmentUri);
         if (segmentByteRangeLength == C.LENGTH_UNSET) {
           // The segment has no byte range defined.
@@ -1480,24 +1533,24 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
     return Long.toHexString(segmentMediaSequence);
   }
 
-  private static @C.SelectionFlags int parseSelectionFlags(String line) {
+  private static @C.SelectionFlags int parseSelectionFlags(String line, MatcherCache matcherCache) {
     int flags = 0;
-    if (parseOptionalBooleanAttribute(line, REGEX_DEFAULT, false)) {
+    if (parseOptionalBooleanAttribute(line, REGEX_DEFAULT, false, matcherCache)) {
       flags |= C.SELECTION_FLAG_DEFAULT;
     }
-    if (parseOptionalBooleanAttribute(line, REGEX_FORCED, false)) {
+    if (parseOptionalBooleanAttribute(line, REGEX_FORCED, false, matcherCache)) {
       flags |= C.SELECTION_FLAG_FORCED;
     }
-    if (parseOptionalBooleanAttribute(line, REGEX_AUTOSELECT, false)) {
+    if (parseOptionalBooleanAttribute(line, REGEX_AUTOSELECT, false, matcherCache)) {
       flags |= C.SELECTION_FLAG_AUTOSELECT;
     }
     return flags;
   }
 
   private static @C.RoleFlags int parseRoleFlags(
-      String line, Map<String, String> variableDefinitions) {
+      String line, Map<String, String> variableDefinitions, MatcherCache matcherCache) {
     String concatenatedCharacteristics =
-        parseOptionalStringAttr(line, REGEX_CHARACTERISTICS, variableDefinitions);
+        parseOptionalStringAttr(line, REGEX_CHARACTERISTICS, variableDefinitions, matcherCache);
     if (TextUtils.isEmpty(concatenatedCharacteristics)) {
       return 0;
     }
@@ -1520,12 +1573,16 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
 
   @Nullable
   private static SchemeData parseDrmSchemeData(
-      String line, String keyFormat, Map<String, String> variableDefinitions)
+      String line,
+      String keyFormat,
+      Map<String, String> variableDefinitions,
+      MatcherCache matcherCache)
       throws ParserException {
     String keyFormatVersions =
-        parseOptionalStringAttr(line, REGEX_KEYFORMATVERSIONS, "1", variableDefinitions);
+        parseOptionalStringAttr(
+            line, REGEX_KEYFORMATVERSIONS, "1", variableDefinitions, matcherCache);
     if (KEYFORMAT_WIDEVINE_PSSH_BINARY.equals(keyFormat)) {
-      String uriString = parseStringAttr(line, REGEX_URI, variableDefinitions);
+      String uriString = parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
       return new SchemeData(
           C.WIDEVINE_UUID,
           MimeTypes.VIDEO_MP4,
@@ -1533,7 +1590,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
     } else if (KEYFORMAT_WIDEVINE_PSSH_JSON.equals(keyFormat)) {
       return new SchemeData(C.WIDEVINE_UUID, "hls", Util.getUtf8Bytes(line));
     } else if (KEYFORMAT_PLAYREADY.equals(keyFormat) && "1".equals(keyFormatVersions)) {
-      String uriString = parseStringAttr(line, REGEX_URI, variableDefinitions);
+      String uriString = parseStringAttr(line, REGEX_URI, variableDefinitions, matcherCache);
       byte[] data = Base64.decode(uriString.substring(uriString.indexOf(',')), Base64.DEFAULT);
       byte[] psshData = PsshAtomUtil.buildPsshAtom(C.PLAYREADY_UUID, data);
       return new SchemeData(C.PLAYREADY_UUID, MimeTypes.VIDEO_MP4, psshData);
@@ -1541,28 +1598,34 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
     return null;
   }
 
-  private static HlsMediaPlaylist.ServerControl parseServerControl(String line) {
+  private static HlsMediaPlaylist.ServerControl parseServerControl(
+      String line, MatcherCache matcherCache) {
     double skipUntilSeconds =
-        parseOptionalDoubleAttr(line, REGEX_CAN_SKIP_UNTIL, /* defaultValue= */ C.TIME_UNSET);
+        parseOptionalDoubleAttr(
+            line, REGEX_CAN_SKIP_UNTIL, /* defaultValue= */ C.TIME_UNSET, matcherCache);
     long skipUntilUs =
         skipUntilSeconds == C.TIME_UNSET
             ? C.TIME_UNSET
             : (long) (skipUntilSeconds * C.MICROS_PER_SECOND);
     boolean canSkipDateRanges =
-        parseOptionalBooleanAttribute(line, REGEX_CAN_SKIP_DATE_RANGES, /* defaultValue= */ false);
+        parseOptionalBooleanAttribute(
+            line, REGEX_CAN_SKIP_DATE_RANGES, /* defaultValue= */ false, matcherCache);
     double holdBackSeconds =
-        parseOptionalDoubleAttr(line, REGEX_HOLD_BACK, /* defaultValue= */ C.TIME_UNSET);
+        parseOptionalDoubleAttr(
+            line, REGEX_HOLD_BACK, /* defaultValue= */ C.TIME_UNSET, matcherCache);
     long holdBackUs =
         holdBackSeconds == C.TIME_UNSET
             ? C.TIME_UNSET
             : (long) (holdBackSeconds * C.MICROS_PER_SECOND);
-    double partHoldBackSeconds = parseOptionalDoubleAttr(line, REGEX_PART_HOLD_BACK, C.TIME_UNSET);
+    double partHoldBackSeconds =
+        parseOptionalDoubleAttr(line, REGEX_PART_HOLD_BACK, C.TIME_UNSET, matcherCache);
     long partHoldBackUs =
         partHoldBackSeconds == C.TIME_UNSET
             ? C.TIME_UNSET
             : (long) (partHoldBackSeconds * C.MICROS_PER_SECOND);
     boolean canBlockReload =
-        parseOptionalBooleanAttribute(line, REGEX_CAN_BLOCK_RELOAD, /* defaultValue= */ false);
+        parseOptionalBooleanAttribute(
+            line, REGEX_CAN_BLOCK_RELOAD, /* defaultValue= */ false, matcherCache);
 
     return new HlsMediaPlaylist.ServerControl(
         skipUntilUs, canSkipDateRanges, holdBackUs, partHoldBackUs, canBlockReload);
@@ -1574,44 +1637,53 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         : C.CENC_TYPE_cbcs;
   }
 
-  private static int parseIntAttr(String line, Pattern pattern) throws ParserException {
-    return Integer.parseInt(parseStringAttr(line, pattern, Collections.emptyMap()));
+  private static int parseIntAttr(String line, Pattern pattern, MatcherCache matcherCache)
+      throws ParserException {
+    return Integer.parseInt(parseStringAttr(line, pattern, Collections.emptyMap(), matcherCache));
   }
 
-  private static int parseOptionalIntAttr(String line, Pattern pattern, int defaultValue) {
-    Matcher matcher = pattern.matcher(line);
+  private static int parseOptionalIntAttr(
+      String line, Pattern pattern, int defaultValue, MatcherCache matcherCache) {
+    Matcher matcher = matcherCache.obtainMatcher(pattern, line);
     if (matcher.find()) {
       return Integer.parseInt(checkNotNull(matcher.group(1)));
     }
     return defaultValue;
   }
 
-  private static long parseLongAttr(String line, Pattern pattern) throws ParserException {
-    return Long.parseLong(parseStringAttr(line, pattern, Collections.emptyMap()));
+  private static long parseLongAttr(String line, Pattern pattern, MatcherCache matcherCache)
+      throws ParserException {
+    return Long.parseLong(parseStringAttr(line, pattern, Collections.emptyMap(), matcherCache));
   }
 
-  private static long parseOptionalLongAttr(String line, Pattern pattern, long defaultValue) {
-    Matcher matcher = pattern.matcher(line);
+  private static long parseOptionalLongAttr(
+      String line, Pattern pattern, long defaultValue, MatcherCache matcherCache) {
+    Matcher matcher = matcherCache.obtainMatcher(pattern, line);
     if (matcher.find()) {
       return Long.parseLong(checkNotNull(matcher.group(1)));
     }
     return defaultValue;
   }
 
-  private static long parseTimeSecondsToUs(String line, Pattern pattern) throws ParserException {
-    String timeValueSeconds = parseStringAttr(line, pattern, Collections.emptyMap());
+  private static long parseTimeSecondsToUs(String line, Pattern pattern, MatcherCache matcherCache)
+      throws ParserException {
+    String timeValueSeconds = parseStringAttr(line, pattern, Collections.emptyMap(), matcherCache);
     BigDecimal timeValue = new BigDecimal(timeValueSeconds);
     return timeValue.multiply(new BigDecimal(C.MICROS_PER_SECOND)).longValue();
   }
 
-  private static double parseDoubleAttr(String line, Pattern pattern) throws ParserException {
-    return Double.parseDouble(parseStringAttr(line, pattern, Collections.emptyMap()));
+  private static double parseDoubleAttr(String line, Pattern pattern, MatcherCache matcherCache)
+      throws ParserException {
+    return Double.parseDouble(parseStringAttr(line, pattern, Collections.emptyMap(), matcherCache));
   }
 
   private static String parseStringAttr(
-      String line, Pattern pattern, Map<String, String> variableDefinitions)
+      String line,
+      Pattern pattern,
+      Map<String, String> variableDefinitions,
+      MatcherCache matcherCache)
       throws ParserException {
-    String value = parseOptionalStringAttr(line, pattern, variableDefinitions);
+    String value = parseOptionalStringAttr(line, pattern, variableDefinitions, matcherCache);
     if (value != null) {
       return value;
     } else {
@@ -1622,24 +1694,29 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
 
   @Nullable
   private static String parseOptionalStringAttr(
-      String line, Pattern pattern, Map<String, String> variableDefinitions) {
-    return parseOptionalStringAttr(line, pattern, null, variableDefinitions);
+      String line,
+      Pattern pattern,
+      Map<String, String> variableDefinitions,
+      MatcherCache matcherCache) {
+    return parseOptionalStringAttr(line, pattern, null, variableDefinitions, matcherCache);
   }
 
   private static @PolyNull String parseOptionalStringAttr(
       String line,
       Pattern pattern,
       @PolyNull String defaultValue,
-      Map<String, String> variableDefinitions) {
-    Matcher matcher = pattern.matcher(line);
+      Map<String, String> variableDefinitions,
+      MatcherCache matcherCache) {
+    Matcher matcher = matcherCache.obtainMatcher(pattern, line);
     @PolyNull String value = matcher.find() ? checkNotNull(matcher.group(1)) : defaultValue;
     return variableDefinitions.isEmpty() || value == null
         ? value
-        : replaceVariableReferences(value, variableDefinitions);
+        : replaceVariableReferences(value, variableDefinitions, matcherCache);
   }
 
-  private static double parseOptionalDoubleAttr(String line, Pattern pattern, double defaultValue) {
-    Matcher matcher = pattern.matcher(line);
+  private static double parseOptionalDoubleAttr(
+      String line, Pattern pattern, double defaultValue, MatcherCache matcherCache) {
+    Matcher matcher = matcherCache.obtainMatcher(pattern, line);
     if (matcher.find()) {
       return Double.parseDouble(checkNotNull(matcher.group(1)));
     }
@@ -1647,7 +1724,10 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
   }
 
   private static HlsMediaPlaylist.ClientDefinedAttribute parseClientDefinedAttribute(
-      String attributes, String clientAttribute, Map<String, String> variableDefinitions)
+      String attributes,
+      String clientAttribute,
+      Map<String, String> variableDefinitions,
+      MatcherCache matcherCache)
       throws ParserException {
     String prefix = clientAttribute + "=";
     int index = attributes.indexOf(prefix);
@@ -1658,27 +1738,26 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
     if (valueBegin.startsWith("\"")) {
       // a quoted string value
       Pattern pattern = Pattern.compile(clientAttribute + "=" + ATTR_QUOTED_STRING_VALUE_PATTERN);
-      String value = parseStringAttr(attributes, pattern, variableDefinitions);
+      String value = parseStringAttr(attributes, pattern, variableDefinitions, matcherCache);
       return new HlsMediaPlaylist.ClientDefinedAttribute(
           clientAttribute, value, HlsMediaPlaylist.ClientDefinedAttribute.TYPE_TEXT);
     } else if (valueBegin.equals("0x") || valueBegin.equals("0X")) {
       // a hexadecimal sequence value
       Pattern pattern = Pattern.compile(clientAttribute + "=(0[xX][A-F0-9]+)");
-      String value = parseStringAttr(attributes, pattern, variableDefinitions);
+      String value = parseStringAttr(attributes, pattern, variableDefinitions, matcherCache);
       return new HlsMediaPlaylist.ClientDefinedAttribute(
           clientAttribute, value, HlsMediaPlaylist.ClientDefinedAttribute.TYPE_HEX_TEXT);
     } else {
       // a decimal-floating-point value
       Pattern pattern = Pattern.compile(clientAttribute + "=([\\d\\.]+)\\b");
       return new HlsMediaPlaylist.ClientDefinedAttribute(
-          clientAttribute, parseDoubleAttr(attributes, pattern));
+          clientAttribute, parseDoubleAttr(attributes, pattern, matcherCache));
     }
   }
 
   private static String replaceVariableReferences(
-      String string, Map<String, String> variableDefinitions) {
-    Matcher matcher = REGEX_VARIABLE_REFERENCE.matcher(string);
-    // TODO: Replace StringBuffer with StringBuilder once Java 9 is available.
+      String string, Map<String, String> variableDefinitions, MatcherCache matcherCache) {
+    Matcher matcher = matcherCache.obtainMatcher(REGEX_VARIABLE_REFERENCE, string);
     StringBuffer stringWithReplacements = new StringBuffer();
     while (matcher.find()) {
       String groupName = matcher.group(1);
@@ -1686,7 +1765,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         matcher.appendReplacement(
             stringWithReplacements, Matcher.quoteReplacement(variableDefinitions.get(groupName)));
       } else {
-        // The variable is not defined. The value is ignored.
+        // The variable is not defined. Leave the reference unchanged.
       }
     }
     matcher.appendTail(stringWithReplacements);
@@ -1702,8 +1781,8 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
   }
 
   private static boolean parseOptionalBooleanAttribute(
-      String line, Pattern pattern, boolean defaultValue) {
-    Matcher matcher = pattern.matcher(line);
+      String line, Pattern pattern, boolean defaultValue, MatcherCache matcherCache) {
+    Matcher matcher = matcherCache.obtainMatcher(pattern, line);
     if (matcher.find()) {
       return BOOLEAN_TRUE.equals(matcher.group(1));
     }
@@ -1753,6 +1832,35 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
       } else {
         throw new NoSuchElementException();
       }
+    }
+  }
+
+  private static final class MatcherCache extends LinkedHashMap<Pattern, Matcher> {
+    private MatcherCache() {
+      super(/* initialCapacity= */ 16, /* loadFactor= */ 0.75f, /* accessOrder= */ true);
+    }
+
+    @Override
+    protected boolean removeEldestEntry(Map.Entry<Pattern, Matcher> eldest) {
+      return size() > 32;
+    }
+
+    /**
+     * Retrieves a cached {@link Matcher} for the given pattern, or creates a new one if not cached.
+     *
+     * @param pattern The regex pattern.
+     * @param input The input to match against.
+     * @return A Matcher for the pattern, reset to match the input.
+     */
+    private Matcher obtainMatcher(Pattern pattern, CharSequence input) {
+      Matcher matcher = this.get(pattern);
+      if (matcher == null) {
+        matcher = pattern.matcher(input);
+        this.put(pattern, matcher);
+      } else {
+        matcher.reset(input);
+      }
+      return matcher;
     }
   }
 }